scripts/ml/requirements.txt
pandas>=2.0.0
pyarrow>=14.0.0
psycopg2-binary>=2.9.9
xgboost>=2.0.0
scikit-learn>=1.4.0
shap>=0.44.0
mlflow>=2.10.0
matplotlib>=3.8.0
numpy>=1.26.0
joblib>=1.3.0
